+++ inc/customizer/class-storefront-customizer.php
@@ -30,6 +30,7 @@ if ( ! class_exists( 'Storefront_Customizer' ) ) :
 			add_action( 'wp_enqueue_scripts',              array( $this, 'add_customizer_css' ), 130 );
 			add_action( 'after_setup_theme',               array( $this, 'custom_header_setup' ) );
 			add_action( 'customize_controls_print_styles', array( $this, 'customizer_custom_control_css' ) );
+			add_action( 'customize_register',              array( $this, 'edit_default_customizer_settings' ), 99 );
 			add_action( 'init',                            array( $this, 'default_theme_mod_values' ), 10 );
 
 			add_action( 'after_switch_theme',              array( $this, 'set_storefront_style_theme_mods' ) );
@@ -73,6 +74,20 @@ if ( ! class_exists( 'Storefront_Customizer' ) ) :
 				add_filter( 'theme_mod_' . $mod, function( $setting ) use ( $val ) {
 					return $setting ? $setting : $val;
 				}, 10 );
+			}
+		}
+
+		/**
+		 * Set Customizer setting defaults.
+		 * These defaults need to be applied separately as child themes can filter storefront_setting_default_values
+		 *
+		 * @param  array $wp_customize the Customizer object.
+		 * @uses   get_storefront_default_setting_values()
+		 * @return void
+		 */
+		public function edit_default_customizer_settings( $wp_customize ) {
+			foreach ( Storefront_Customizer::get_storefront_default_setting_values() as $mod => $val ) {
+				$wp_customize->get_setting( $mod )->default = $val;
 			}
 		}